pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qless-solver"
version = "0.1.0"
description = "A solver for the Q-Less solitaire dice game"
readme = "README.md"
authors = [
    {name = "Scott Pritchard"}
]
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
dependencies = [
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "jinja2>=3.0.0",
    "pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.0.250",
    "httpx>=0.23.0",
]

[project.scripts]
qless-solver = "qless_solver.cli:main"

[tool.setuptools]
package-dir = {"" = "cli"}
packages = ["qless_solver"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
files = ["cli"]
exclude = "(?x)(tests/|web/)"

[tool.ruff]
line-length = 88
target-version = "py312"
